chore: Organize some function better.
